+++ src/solarPower.test.ts
@@ -50,33 +50,6 @@ if (!debug) {
   consoleErrorSpy = jest.spyOn(console, 'error');
 }
 
-/**
- * Waits for the `isOnline` property to become `true`.
- * This function checks the `isOnline` property of the provided server node at regular intervals until it becomes `true` or the specified timeout is reached.
- * If the timeout is reached before `isOnline` becomes `true`, the promise is rejected with an error.
- *
- * @param {ServerNode<ServerNode.RootEndpoint>} server - The server node to check for online status.
- * @param {number} timeout - The maximum time to wait in milliseconds.
- * @returns {Promise<void>} A promise that resolves when `isOnline` becomes `true` or rejects if the timeout is reached.
- */
-async function waitForOnline(server: ServerNode<ServerNode.RootEndpoint>, timeout = 10000): Promise<void> {
-  const start = Date.now();
-
-  return new Promise((resolve, reject) => {
-    const checkOnline = () => {
-      if (server.lifecycle.isOnline) {
-        resolve();
-      } else if (Date.now() - start >= timeout) {
-        reject(new Error('Timeout waiting for server.lifecycle.isOnline to become true'));
-      } else {
-        setTimeout(checkOnline, 100); // Check every 100ms
-      }
-    };
-    // Start checking immediately
-    checkOnline();
-  });
-}
-
 // Cleanup the matter environment
 rmSync(HOMEDIR, { recursive: true, force: true });
 
@@ -180,11 +153,20 @@ describe('Matterbridge ' + NAME, () => {
 
   test('start the server node', async () => {
     // Run the server
-    await server.start();
+    expect(server.lifecycle.isReady).toBeTruthy();
+    expect(server.lifecycle.isOnline).toBeFalsy();
+
+    // Wait for the server to be online
+    await new Promise<void>((resolve) => {
+      server.lifecycle.online.on(async () => {
+        resolve();
+      });
+      server.start();
+    });
+
+    // Check if the server is online
     expect(server.lifecycle.isReady).toBeTruthy();
     expect(server.lifecycle.isOnline).toBeTruthy();
-    // Wait for the server to be online
-    await waitForOnline(server);
   });
 
   test('device forEachAttribute', async () => {